cv-review-agent--main/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-core
langchain-openai
langchain-google-genai
langgraph
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
pydantic[email]
PyPDF2
python-docx
pypdf
schedule